+++ CHANGELOG.md
@@ -1,6 +1,76 @@
 # Change Log
 All notable changes to the "ms-vscode.azure-account" extension will be documented in this file.
 
+## [0.12.0] - 2024-05-14
+
+In preparation of the [Azure Account extension being deprecated](https://github.com/microsoft/vscode-azure-account/issues/964) at the end of the year, we've moved the Azure Cloud Shell feature to the [Azure Resources extension](https://marketplace.visualstudio.com/items?itemName=ms-azuretools.vscode-azureresourcegroups). Apart from moving codebases, the feature is the same from a users perspective. Authentication for the Cloud Shell feature is now handled by the VS Code built-in Microsoft authentication provider, which means you may have to login upon first use of the migrated feature.
+
+Additionally, we've fixed two longstanding Azure Cloud Shell bugs that caused issues launching the feature on Linux and macOS: [#719](https://github.com/microsoft/vscode-azure-account/issues/719) and [#959](https://github.com/microsoft/vscode-azure-account/issues/959).
+
+### Changed
+* [[958]](https://github.com/microsoft/vscode-azure-account/pull/958) Depend on Azure Resources extension for the Azure Cloud Shell feature 
+
+### Fixed
+
+* [[855]](https://github.com/microsoft/vscode-azureresourcegroups/pull/855) Stop using `--ms-enable-electron-run-as-node` flag to fix launching Cloud Shell on macOS
+* [[854]](https://github.com/microsoft/vscode-azureresourcegroups/pull/854) Use `process.execPath` instead of `process.argv0` to fix launching Cloud Shell on Linux
+
+## [0.11.7] - 2024-04-30
+
+### Added
+* Support ephemeral cloud shell sessions [#950](https://github.com/microsoft/vscode-azure-account/pull/950)
+
+## [0.11.6] - 2023-10-06
+
+### Fixed
+* Fix 500 error when launching cloud shell by setting the referer header [#866](https://github.com/microsoft/vscode-azure-account/pull/866)
+
+### Changed
+* Add manage account command to make sign out easier to find in [#820](https://github.com/microsoft/vscode-azure-account/pull/820)
+* Remove use of AAD Graph for cloud shell in [#851](https://github.com/microsoft/vscode-azure-account/pull/851)
+
+## [0.11.5] - 2023-05-16
+
+### Fixed
+* Fix launching cloud console by @alexweininger in [#809](https://github.com/microsoft/vscode-azure-account/pull/809)
+
+## [0.11.4] - 2023-05-02
+
+### Added
+* Add detailed logging  by @alexweininger in [#750](https://github.com/microsoft/vscode-azure-account/pull/750)
+
+### Fixed
+* Don't await nps survey by @alexweininger in [#709](https://github.com/microsoft/vscode-azure-account/pull/709)
+
+### Dependencies
+* Bump version post release by @alexweininger in [#701](https://github.com/microsoft/vscode-azure-account/pull/701)
+* Bump webpack from 5.69.0 to 5.76.0 by @dependabot in [#736](https://github.com/microsoft/vscode-azure-account/pull/736)
+* Bump xml2js and @azure/ms-rest-js in /sample by @dependabot in [#778](https://github.com/microsoft/vscode-azure-account/pull/778)
+
+## [0.11.3] - 2023-01-18
+
+### Fixed
+* Fixed issues with Azure Cloud Shell terminal when connected to a remote host @alexweininger in https://github.com/microsoft/vscode-azure-account/pull/684
+
+### Dependencies
+* Bump @xmldom/xmldom from 0.7.5 to 0.7.8 by @dependabot in [#663](https://github.com/microsoft/vscode-azure-account/pull/663)
+* Bump loader-utils from 1.4.0 to 1.4.1 by @dependabot in [#666](https://github.com/microsoft/vscode-azure-account/pull/666)
+* Bump loader-utils from 1.4.1 to 1.4.2 by @dependabot in [#670](https://github.com/microsoft/vscode-azure-account/pull/670)
+* Bump decode-uri-component from 0.2.0 to 0.2.2 by @dependabot in [#678](https://github.com/microsoft/vscode-azure-account/pull/678)
+* Bump json5 from 1.0.1 to 1.0.2 by @dependabot in [#692](https://github.com/microsoft/vscode-azure-account/pull/692)
+* Bump jsonwebtoken and @azure/msal-node by @dependabot in [#696](https://github.com/microsoft/vscode-azure-account/pull/696)
+
+### Other
+* Fix readme badges by @bwateratmsft in [#653](https://github.com/microsoft/vscode-azure-account/pull/653)
+
+## [0.11.2] - 2022-10-07
+
+### Added
+- Add Support for Workspace Trust [#624](https://github.com/microsoft/vscode-azure-account/pull/624)
+
+### Fixed
+- Update http-proxy-agent and https-proxy-agent [#640](https://github.com/microsoft/vscode-azure-account/pull/640)
+
 ## [0.11.1] - 2022-08-03
 
 ### Fixed